requirements.txt
# requirements.txt
# Python dependencies for the Personal Learning Assistant app.
# These packages are required for PDF processing, LLM integration, embeddings, vector search, and environment management.

streamlit
langchain>=0.2.12
langchain-community>=0.2.11
openai>=1.42.0
pypdf>=4.3.1
chromadb>=0.5.5
langchain-openai
python-dotenv
# - streamlit: Web app framework
# - langchain, langchain-community, langchain-openai: LLM and embeddings integration
# - openai: OpenAI API client
# - pypdf: PDF reading
# - chromadb: Vector store for similarity search
# - python-dotenv: Environment variable management